CsvImporters: Extract CSV parsing code into GenericCSVParser. Update ReferenceCsvImport and AltiumCsvImporter to use extracted code.  Update DipTraceBoardImporter to use the same pattern.
* This goal is so that the DipTraceBoardImporter can be updated to use the GenericCSVParser.
