+++ src/main/java/org/openpnp/gui/importer/altium/AltiumCsvImporter.java
@@ -26,11 +26,14 @@
  * column headers generated by Altum, making it a little bit more robust.
  */
 
-package org.openpnp.gui.importer;
+package org.openpnp.gui.importer.altium;
 
 import java.awt.Frame;
 
 import org.openpnp.Translations;
+import org.openpnp.gui.importer.BoardImporter;
+import org.openpnp.gui.importer.genericcsv.CsvImporter;
+import org.openpnp.gui.importer.genericcsv.csv.GenericCSVParser;
 import org.openpnp.model.Abstract2DLocatable.Side;
 import org.openpnp.model.Board;
 import org.openpnp.model.Location;
@@ -39,8 +42,6 @@ import org.openpnp.model.Placement;
 public class AltiumCsvImporter extends CsvImporter implements BoardImporter {
     private final static String NAME = "Altium .csv"; //$NON-NLS-1$
     private final static String DESCRIPTION = Translations.getString("AltiumCsvImporter.Importer.Description"); //$NON-NLS-1$
-
-    private Board board;
 
     @Override
     public String getImporterName() {
@@ -55,7 +56,7 @@ public class AltiumCsvImporter extends CsvImporter implements BoardImporter {
     @Override
     public Board importBoard(Frame parent) throws Exception {
     	// run Named CSV importer
-    	board = super.importBoard(parent);
+    	Board board = super.importBoard(parent, this);
     	
     	// modify rotation of bottom side components
     	if (board != null) {
@@ -71,59 +72,66 @@ public class AltiumCsvImporter extends CsvImporter implements BoardImporter {
     	return board;
     }
 
-    // methods to read of strings for each purpose to be found in the heading line
-    // data read from file is converted to upper case before compare -> only list upper case pattern here
-    // this is the list that is used by the default Named CSV importer
-    public String[] getReferencePattern() {
-    	return new String[] {
-    			"DESIGNATOR" 	//$NON-NLS-1$
-    	};
-    }
-    public String[] getValuePattern() {
-    	return new String[] {
-    			"COMMENT"	 	//$NON-NLS-1$
-    	};
-    }
-    public String[] getPackagePattern() {
-    	return new String[] {
-    			"FOOTPRINT"	 	//$NON-NLS-1$
-    	};
-    }
-    public String[] getXPattern() {
-    	return new String[] {
-    			"REF-X(MM)",  	//$NON-NLS-1$
-    			"REF-X(MIL)",  	//$NON-NLS-1$
-    			"CENTER-X(MM)", //$NON-NLS-1$
-    			"CENTER-X(MIL)"	//$NON-NLS-1$
-    	};
-    }
-    public String[] getYPattern() {
-    	return new String[] {
-    			"REF-Y(MM)",  	//$NON-NLS-1$
-    			"REF-Y(MIL)",  	//$NON-NLS-1$
-    			"CENTER-Y(MM)", //$NON-NLS-1$
-    			"CENTER-Y(MIL)"	//$NON-NLS-1$
-    	};
-    }
-    public String[] getRotationPattern() {
-    	return new String[] {
-    		    "ROTATION" 		//$NON-NLS-1$
-    	};
-    }
-    public String[] getSidePattern() {
-    	return new String[] {
-    		    "LAYER"			//$NON-NLS-1$
-    	};
-    }
-    public String[] getHeightPattern() {
-    	return new String[] {
-    		    "HEIGHT(MIL)",	//$NON-NLS-1$ 
-    		    "HEIGHT(MM)"  	//$NON-NLS-1$
-    	};
-    }
-    public String[] getCommentPattern() {
-    	return new String[] {
-    		    "DESCRIPTION" 	//$NON-NLS-1$
-    	};
-    }
+	@Override
+	public GenericCSVParser getParser() {
+		return new AltiumCSVParser();
+	}
+
+	private class AltiumCSVParser extends GenericCSVParser {
+		// methods to read of strings for each purpose to be found in the heading line
+		// data read from file is converted to upper case before compare -> only list upper case pattern here
+		// this is the list that is used by the default Named CSV importer
+		public String[] getReferencePattern() {
+			return new String[] {
+					"DESIGNATOR" 	//$NON-NLS-1$
+			};
+		}
+		public String[] getValuePattern() {
+			return new String[] {
+					"COMMENT"	 	//$NON-NLS-1$
+			};
+		}
+		public String[] getPackagePattern() {
+			return new String[] {
+					"FOOTPRINT"	 	//$NON-NLS-1$
+			};
+		}
+		public String[] getXPattern() {
+			return new String[] {
+					"REF-X(MM)",  	//$NON-NLS-1$
+					"REF-X(MIL)",  	//$NON-NLS-1$
+					"CENTER-X(MM)", //$NON-NLS-1$
+					"CENTER-X(MIL)"	//$NON-NLS-1$
+			};
+		}
+		public String[] getYPattern() {
+			return new String[] {
+					"REF-Y(MM)",  	//$NON-NLS-1$
+					"REF-Y(MIL)",  	//$NON-NLS-1$
+					"CENTER-Y(MM)", //$NON-NLS-1$
+					"CENTER-Y(MIL)"	//$NON-NLS-1$
+			};
+		}
+		public String[] getRotationPattern() {
+			return new String[] {
+					"ROTATION" 		//$NON-NLS-1$
+			};
+		}
+		public String[] getSidePattern() {
+			return new String[] {
+					"LAYER"			//$NON-NLS-1$
+			};
+		}
+		public String[] getHeightPattern() {
+			return new String[] {
+					"HEIGHT(MIL)",	//$NON-NLS-1$
+					"HEIGHT(MM)"  	//$NON-NLS-1$
+			};
+		}
+		public String[] getCommentPattern() {
+			return new String[] {
+					"DESCRIPTION" 	//$NON-NLS-1$
+			};
+		}
+	}
 }
