+++ src/main/java/org/openpnp/gui/importer/genericcsv/gui/Dlg.java
@@ -1,14 +1,15 @@
-package org.openpnp.gui.importer.diptrace.gui;
+package org.openpnp.gui.importer.genericcsv.gui;
 
 import com.jgoodies.forms.layout.ColumnSpec;
 import com.jgoodies.forms.layout.FormLayout;
 import com.jgoodies.forms.layout.FormSpecs;
 import com.jgoodies.forms.layout.RowSpec;
 import org.openpnp.Translations;
-import org.openpnp.gui.importer.diptrace.csv.DipTraceCSVParser;
+import org.openpnp.gui.importer.genericcsv.csv.GenericCSVParser;
 import org.openpnp.gui.support.MessageBoxes;
 import org.openpnp.model.Board;
 import org.openpnp.model.Placement;
+import org.pmw.tinylog.Logger;
 
 import javax.swing.*;
 import javax.swing.border.TitledBorder;
@@ -19,33 +20,25 @@ import java.io.FilenameFilter;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DiptraceBoardImporterDialog extends JDialog {
+public class Dlg extends JDialog {
+    private final GenericCSVParser parser;
 
-    private final DipTraceCSVParser parser;
-    private File fileName;
-
-    public Board getBoard() {
-        return board;
-    }
-
-    private Board board;
-
-    //, bottomFile;
-
-    private JTextField textFieldFileName;
-    private JTextField textFieldBottomFile;
-    private final Action browseTopFileAction = new SwingAction();
+    private JTextField textFieldFile;
+    private final Action browseFileAction = new SwingAction();
     private final Action importAction = new SwingAction_2();
     private final Action cancelAction = new SwingAction_3();
     private JCheckBox chckbxCreateMissingParts;
+    private JCheckBox chckbxUpdatePartHeight;
+    private Board board;
 
-    public DiptraceBoardImporterDialog(Frame parent, String title, DipTraceCSVParser parser) {
+    public Dlg(Frame parent, String title, GenericCSVParser parser) {
         super(parent, title, true);
+
         this.parser = parser;
         getContentPane().setLayout(new BoxLayout(getContentPane(), BoxLayout.Y_AXIS));
 
         JPanel panel = new JPanel();
-        panel.setBorder(new TitledBorder(null, Translations.getString("DipTraceImporter.FilesPanel.Border.title"), TitledBorder.LEADING, TitledBorder.TOP, //$NON-NLS-1$
+        panel.setBorder(new TitledBorder(null, Translations.getString("CsvImporter.FilesPanel.Border.title"), TitledBorder.LEADING, TitledBorder.TOP, //$NON-NLS-1$
                 null, null));
         getContentPane().add(panel);
         panel.setLayout(new FormLayout(
@@ -55,28 +48,33 @@ public class DiptraceBoardImporterDialog extends JDialog {
                 new RowSpec[]{FormSpecs.RELATED_GAP_ROWSPEC, FormSpecs.DEFAULT_ROWSPEC,
                         FormSpecs.RELATED_GAP_ROWSPEC, FormSpecs.DEFAULT_ROWSPEC,}));
 
-        JLabel lblTopFilemnt = new JLabel(Translations.getString("DipTraceImporter.FilesPanel.TopFilemntLabel.text")); //$NON-NLS-1$
+        JLabel lblTopFilemnt = new JLabel(Translations.getString("CsvImporter.FilesPanel.topFilemntLabel.text")); //$NON-NLS-1$
         panel.add(lblTopFilemnt, "2, 2, right, default"); //$NON-NLS-1$
 
-        textFieldFileName = new JTextField();
-        panel.add(textFieldFileName, "4, 2, fill, default"); //$NON-NLS-1$
-        textFieldFileName.setColumns(10);
+        textFieldFile = new JTextField();
+        panel.add(textFieldFile, "4, 2, fill, default"); //$NON-NLS-1$
+        textFieldFile.setColumns(30);
 
-        JButton btnBrowse = new JButton(Translations.getString("DipTraceImporter.FilesPanel.browseButton.text")); //$NON-NLS-1$
-        btnBrowse.setAction(browseTopFileAction);
+        JButton btnBrowse = new JButton(Translations.getString("CsvImporter.FilesPanel.browseButton.text")); //$NON-NLS-1$
+        btnBrowse.setAction(browseFileAction);
         panel.add(btnBrowse, "6, 2"); //$NON-NLS-1$
 
         JPanel panel_1 = new JPanel();
-        panel_1.setBorder(new TitledBorder(null, Translations.getString("DipTraceImporter.OptionsPanel.Border.title"), TitledBorder.LEADING, //$NON-NLS-1$
+        panel_1.setBorder(new TitledBorder(null, Translations.getString("CsvImporter.OptionsPanel.Border.title"), TitledBorder.LEADING, //$NON-NLS-1$
                 TitledBorder.TOP, null, null));
         getContentPane().add(panel_1);
         panel_1.setLayout(new FormLayout(
                 new ColumnSpec[]{FormSpecs.RELATED_GAP_COLSPEC, FormSpecs.DEFAULT_COLSPEC,},
-                new RowSpec[]{FormSpecs.RELATED_GAP_ROWSPEC, FormSpecs.DEFAULT_ROWSPEC,}));
+                new RowSpec[]{FormSpecs.RELATED_GAP_ROWSPEC, FormSpecs.DEFAULT_ROWSPEC,
+                        RowSpec.decode("default:grow")})); //$NON-NLS-1$
 
-        chckbxCreateMissingParts = new JCheckBox(Translations.getString("DipTraceImporter.OptionsPanel.createMissingPartsCheckbox.text")); //$NON-NLS-1$
+        chckbxCreateMissingParts = new JCheckBox(Translations.getString("CsvImporter.OptionsPanel.createMissingPartsChkbox.text")); //$NON-NLS-1$
         chckbxCreateMissingParts.setSelected(true);
         panel_1.add(chckbxCreateMissingParts, "2, 2"); //$NON-NLS-1$
+
+        chckbxUpdatePartHeight = new JCheckBox(Translations.getString("CsvImporter.OptionsPanel.updatePartHeightChkbox.text")); //$NON-NLS-1$
+        chckbxUpdatePartHeight.setSelected(true);
+        panel_1.add(chckbxUpdatePartHeight, "2, 3"); //$NON-NLS-1$
 
         JSeparator separator = new JSeparator();
         getContentPane().add(separator);
@@ -86,15 +84,16 @@ public class DiptraceBoardImporterDialog extends JDialog {
         flowLayout.setAlignment(FlowLayout.RIGHT);
         getContentPane().add(panel_2);
 
-        JButton btnCancel = new JButton(Translations.getString("DipTraceImporter.ButtonsPanel.cancelButton.text")); //$NON-NLS-1$
+        JButton btnCancel = new JButton(Translations.getString("CsvImporter.ButtonsPanel.cancelButton.text")); //$NON-NLS-1$
         btnCancel.setAction(cancelAction);
         panel_2.add(btnCancel);
 
-        JButton btnImport = new JButton(Translations.getString("DipTraceImporter.ButtonsPanel.importButton.text")); //$NON-NLS-1$
+        JButton btnImport = new JButton(Translations.getString("CsvImporter.ButtonsPanel.importButton.text")); //$NON-NLS-1$
         btnImport.setAction(importAction);
         panel_2.add(btnImport);
 
-        setSize(400, 400);
+        // resize to the window to its preferred size
+        pack();
         setLocationRelativeTo(parent);
 
         JRootPane rootPane = getRootPane();
@@ -104,18 +103,24 @@ public class DiptraceBoardImporterDialog extends JDialog {
         rootPane.getActionMap().put("ESCAPE", cancelAction); //$NON-NLS-1$
     }
 
+    public Board getBoard() {
+        return board;
+    }
+
     private class SwingAction extends AbstractAction {
         public SwingAction() {
-            putValue(NAME, Translations.getString("DipTraceImporter.BrowseAction.Name")); //$NON-NLS-1$
-            putValue(SHORT_DESCRIPTION, Translations.getString("DipTraceImporter.BrowseAction.ShortDescription")); //$NON-NLS-1$
+            putValue(NAME, Translations.getString("CsvImporter.BrowseAction.Name")); //$NON-NLS-1$
+            putValue(SHORT_DESCRIPTION, Translations.getString("CsvImporter.BrowseAction.ShortDescription")); //$NON-NLS-1$
         }
 
         public void actionPerformed(ActionEvent e) {
-            FileDialog fileDialog = new FileDialog(DiptraceBoardImporterDialog.this);
+            FileDialog fileDialog = new FileDialog(Dlg.this);
             fileDialog.setFilenameFilter(new FilenameFilter() {
                 @Override
                 public boolean accept(File dir, String name) {
-                    return name.toLowerCase().endsWith(".csv"); //$NON-NLS-1$
+                    return false || name.toLowerCase().endsWith(".csv") //$NON-NLS-1$
+                            || name.toLowerCase().endsWith(".txt") //$NON-NLS-1$
+                            || name.toLowerCase().endsWith(".dat"); //$NON-NLS-1$
                 }
             });
             fileDialog.setVisible(true);
@@ -123,29 +128,29 @@ public class DiptraceBoardImporterDialog extends JDialog {
                 return;
             }
             File file = new File(new File(fileDialog.getDirectory()), fileDialog.getFile());
-            textFieldFileName.setText(file.getAbsolutePath());
+            textFieldFile.setText(file.getAbsolutePath());
         }
     }
 
+
     private class SwingAction_2 extends AbstractAction {
         public SwingAction_2() {
-            putValue(NAME, Translations.getString("DipTraceImporter.ImportAction.Name")); //$NON-NLS-1$
-            putValue(SHORT_DESCRIPTION, Translations.getString("DipTraceImporter.ImportAction.ShortDescription")); //$NON-NLS-1$
+            putValue(NAME, Translations.getString("CsvImporter.Import2Action.Name")); //$NON-NLS-1$
+            putValue(SHORT_DESCRIPTION, Translations.getString("CsvImporter.Import2Action.ShortDescription")); //$NON-NLS-1$
         }
 
         public void actionPerformed(ActionEvent e) {
-            fileName = new File(textFieldFileName.getText());
+            Logger.debug("Parsing " + textFieldFile.getText() + " CSV FIle"); //$NON-NLS-1$ //$NON-NLS-2$
+            File file = new File(textFieldFile.getText());
             board = new Board();
             List<Placement> placements = new ArrayList<>();
             try {
-                if (fileName.exists()) {
-                    placements.addAll(parser.parseFile(fileName, chckbxCreateMissingParts.isSelected()));
-
+                if (file.exists()) {
+                    placements.addAll(parser.parseFile(file, chckbxCreateMissingParts.isSelected(),
+                            chckbxUpdatePartHeight.isSelected()));
                 }
             } catch (Exception e1) {
-                MessageBoxes.errorBox(DiptraceBoardImporterDialog.this, "Import Error", "The expected file format is the default file export in DipTrace " //$NON-NLS-1$ //$NON-NLS-2$
-                        + "PCB: File -> Export -> Pick and Place. The first line indicates RefDes, Name, X (mm), Y (mm), Side, Rotate, Value." //$NON-NLS-1$
-                        + "The lines that follow are data."); //$NON-NLS-1$
+                MessageBoxes.errorBox(Dlg.this, Translations.getString("CsvImporter.ImportErrorMessage"), e1); //$NON-NLS-1$
                 return;
             }
             for (Placement placement : placements) {
@@ -157,8 +162,8 @@ public class DiptraceBoardImporterDialog extends JDialog {
 
     private class SwingAction_3 extends AbstractAction {
         public SwingAction_3() {
-            putValue(NAME, Translations.getString("DipTraceImporter.CancelAction.Name")); //$NON-NLS-1$
-            putValue(SHORT_DESCRIPTION, Translations.getString("DipTraceImporter.CancelAction.ShortDescription")); //$NON-NLS-1$
+            putValue(NAME, Translations.getString("CsvImporter.CancelAction.Name")); //$NON-NLS-1$
+            putValue(SHORT_DESCRIPTION, Translations.getString("CsvImporter.CancelAction.ShortDescription")); //$NON-NLS-1$
         }
 
         public void actionPerformed(ActionEvent e) {
